Spec: Add optional `sessionId` field to `serverInfo` message (#360)
**Public-Facing Changes**
- Adds optional `sessionId` field to `serverInfo` message

**Description**
From #256:

> Including sessionId sent by server on opening connection allows client
to understand context and if the connection is a re-connection from a
dropped context, or if it is a new session connecting. We will then be
able to clear state in the client when connecting to a new context, but
retain state when reconnecting a dropped connection.
>
> Context: Customer is currently reconnecting to a websocket in a
different context and the state is being preserved leading to incorrect
results. They would like the state to be cleared when reconnecting to
the WS in a new context.


Fixes #256

Linear:
[FG-1654](https://linear.app/foxglove/issue/FG-1654/include-session-id-in-websocket-connection-%5Bfoxglovews-protocol-256%5D)