Clean up specification (#371)
**Public-Facing Changes**
None

**Description**

From #370:

> * [x] If you click on the link to "Message Data" in the TOC, it takes
you to the client message data, not server message data.
> * [x] Additionally, the client message data op is listed under a
heading called "JSON Messages". The heading called "Binary Messages"
doesn't include the client message data.
> * [x] It looks like Service Call Request and Time have the same opcode
(0x02).

It looks like that because it is currently not obvoius which messages
are sent by the server and which by the client. In the TOC we divide
between server and client messages, but the rest of the document does
not since there it is divided into JSON and binary messages. IMO, the
rest of the document and the heading order should be in the same order
as the TOC.

> * [x] The Service Call Request heading is too small (wrong level)



Fixes #370
Fixes FG-2011