Add parameter type `float64` & `float64_array` for floating point values
Some languages, such as JavaScript, do not differentiate between
integers and floating point numbers. The `float64` and `float64_array`
types were added to avoid ambiguity by informing the client (or the
server) that the value is a floating point number, or an array of it,
respectively.

Example: When sending the following JSON via JS to the server, the
server will receive the value `1` and deduce that the parameter
is of type Integer.

```
{
  "op": "setParameters",
  "parameters": [
    { "name": "/float_param", "value": 1.0 }
  ]
}
```

By adding `"type": "float64"`, the client can let the server know that
the value should be treated as a 64-bit floating point number.