mcap-record: Fix race condition when adding messages (#598)
### Public-Facing Changes

mcap-record: Fix race condition when adding messages

### Description
Does not directly call `writer.addMessage` in the websocket message
handler but instead pushes it to a queue. This makes sure that we do not
call `writer.addMessage` concurrently which could happen due to a
certain `addMessage` taking longer when a chunk is ready to be
compressed & written. This PR also allows to set the message queue size
and to opt out of chunk compression.

Fixes #592
Resolves FG-5775