mcap-record: Avoid duplicate subscription to channel (#744)
### Changelog
Fix `mcap-record` creating duplicated channels if existing channels are
re-advertised by the server

### Docs
None

### Description
Logs a message when the server erroneously re-advertises channels which
have been advertised before. As per the
[spec](https://github.com/foxglove/ws-protocol/blob/main/docs/spec.md#advertise),
only new channels should be advertised by the server but unfortunately
not all servers comply with that. This extra check makes sure that
channels are not duplicated in the resulting MCAP file.