cpp: specify param type for double / double array parameters (#921)
### Changelog
cpp: Add parameter type float64 & float64_array for floating point
values

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description
From https://github.com/foxglove/ws-protocol/pull/519:

> Some languages, such as JavaScript, do not differentiate between
integers and floating point numbers. The `float64` and `float64_array`
types were added to avoid ambiguity by informing the client (or the
server) that the value is a floating point number, or an array of it,
respectively.

On the foxglove_bridge side this has been implemented in
https://github.com/foxglove/ros-foxglove-bridge/pull/256 but that change
has apparently never found its way to this repo. This PR applies that
change to this repo.