Add deprecation notices (#940)
This adds deprecation notices to the README, and to C++ and Python
packages, which are being replaced by the Foxglove SDK
(https://github.com/foxglove/foxglove-sdk).

This bumps the version numbers of C++ and Python packages, so that we
can publish a patch release with deprecation notices on those
repositories.